requirements.txt
# Core dependencies - Python 3.13 compatible
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Data processing
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
colorama>=0.4.6
websocket-client>=1.6.0

# Technical analysis
ta>=0.10.0
scipy>=1.12.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
